Finish prototype, fix major bugs
While the naming logic has not been thoroughly tested, the only
functionality yet to be implemented is chord playback. The mute
button is working well, as well as the refreshing (requesting) of
the results section. Askama was difficult to get started with,
however it has made templating much easier. Furthermore, I have
realized that I have a lot to learn about JSON.